refactor(course): refactoring course card types
