+++ src/app/api/courses.api.slice.ts
@@ -1,4 +1,5 @@
 import { authApi } from 'app/base/auth.api';
+import { setCourses } from 'app/slice/courses.slice';
 
 export const coursesApiSlice = authApi.injectEndpoints({
   endpoints: (builder) => ({
@@ -23,6 +24,10 @@ export const coursesApiSlice = authApi.injectEndpoints({
         url: '/courses/',
         method: 'GET',
       }),
+      async onQueryStarted(_, { dispatch, queryFulfilled }) {
+        const { data } = await queryFulfilled;
+        dispatch(setCourses(data));
+      },
       providesTags: ['Course'],
     }),
     FilterCourses: builder.mutation({
@@ -46,6 +51,7 @@ export const coursesApiSlice = authApi.injectEndpoints({
         url: `/courses/${courseId}`,
         method: 'GET',
       }),
+      providesTags: ['Course'],
     }),
     GetInstructorCourses: builder.query({
       query: () => ({
