+++ src/app/slice/enrolled.slice.ts
@@ -1,13 +1,19 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { EnrolledCourses } from 'shared/ts/types';
 
-const enrolled: EnrolledCourses[] = [];
+const getInitialState = (): EnrolledCourses[] => {
+  const storedCourses = sessionStorage.getItem('enrolled');
+  return storedCourses ? JSON.parse(storedCourses) : [];
+};
+
+const initialState: EnrolledCourses[] = getInitialState();
 
 const enrolledSlice = createSlice({
   name: 'enrolled',
-  initialState: enrolled,
+  initialState,
   reducers: {
     setEnrolled: (_state, action) => {
+      sessionStorage.setItem('enrolled', JSON.stringify(action.payload));
       return action.payload;
     },
   },
